+++ concierge-aerospike/src/main/java/com/grookage/concierge/aerospike/client/AerospikeManager.java
@@ -1,7 +1,7 @@
 package com.grookage.concierge.aerospike.client;
 
-import com.aerospike.client.*;
 import com.aerospike.client.Record;
+import com.aerospike.client.*;
 import com.aerospike.client.exp.Exp;
 import com.aerospike.client.policy.RecordExistsAction;
 import com.aerospike.client.policy.WritePolicy;
@@ -9,6 +9,7 @@ import com.aerospike.client.query.Statement;
 import com.grookage.concierge.aerospike.storage.AerospikeRecord;
 import com.grookage.concierge.aerospike.storage.AerospikeStorageConstants;
 import com.grookage.concierge.models.MapperUtils;
+import java.util.Set;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -84,9 +85,9 @@ public class AerospikeManager {
     }
 
     @SneakyThrows
-    public List<AerospikeRecord> getRecords(final List<String> namespaces,
-                                            final List<String> configNames,
-                                            final List<String> configStates) {
+    public List<AerospikeRecord> getRecords(final Set<String> namespaces,
+                                            final Set<String> configNames,
+                                            final Set<String> configStates) {
         final var queryStatement = new Statement();
         queryStatement.setNamespace(namespace);
         queryStatement.setSetName(AerospikeStorageConstants.CONFIG_SET);