+++ concierge-client/src/main/java/com/grookage/concierge/client/refresher/ConciergeClientSupplier.java
@@ -19,7 +19,8 @@ package com.grookage.concierge.client.refresher;
 import com.google.common.base.Strings;
 import com.google.common.net.HttpHeaders;
 import com.grookage.concierge.models.MapperUtils;
-import com.grookage.concierge.models.NamespaceRequest;
+import com.grookage.concierge.models.SearchRequest;
+import com.grookage.concierge.models.config.ConfigState;
 import com.grookage.concierge.models.ingestion.ConfigurationResponse;
 import com.grookage.leia.provider.config.LeiaHttpConfiguration;
 import com.grookage.leia.provider.suppliers.LeiaHttpSupplier;
@@ -51,7 +52,7 @@ public class ConciergeClientSupplier extends LeiaHttpSupplier<List<Configuration
 
     @Override
     protected String url() {
-        return "/v1/configs/details/active";
+        return "/v1/configs/details";
     }
 
     @Override
@@ -59,8 +60,9 @@ public class ConciergeClientSupplier extends LeiaHttpSupplier<List<Configuration
     protected Request getRequest(String url) {
         final var requestBody = RequestBody.create(
                 okhttp3.MediaType.parse("application/json; charset=utf-8"),
-                MapperUtils.mapper().writeValueAsString(NamespaceRequest.builder()
+                MapperUtils.mapper().writeValueAsString(SearchRequest.builder()
                         .namespaces(namespaces)
+                        .configStates(Set.of(ConfigState.ACTIVATED))
                         .build()));
         final var requestBuilder = new Request.Builder()
                 .url(endPoint(url))