+++ concierge-core/src/main/java/com/grookage/concierge/core/engine/ConciergeHub.java
@@ -5,7 +5,7 @@ import com.grookage.concierge.core.engine.processors.*;
 import com.grookage.concierge.core.engine.resolver.AppendConfigResolver;
 import com.grookage.concierge.core.engine.resolver.ConfigVersionManager;
 import com.grookage.concierge.core.engine.resolver.DefaultAppendConfigResolver;
-import com.grookage.concierge.core.managers.VersionGenerator;
+import com.grookage.concierge.core.managers.ProcessorFactory;
 import com.grookage.concierge.models.config.ConfigEvent;
 import com.grookage.concierge.models.config.ConfigEventVisitor;
 import com.grookage.concierge.repository.ConciergeRepository;
@@ -22,9 +22,9 @@ public class ConciergeHub {
 
     private static final Map<ConfigEvent, ConciergeProcessor> processors = new ConcurrentHashMap<>();
     private Supplier<ConciergeRepository> repositorySupplier;
-    private Supplier<VersionGenerator> versionSupplier;
     private Supplier<AppendConfigResolver> appendConfigResolverSupplier = DefaultAppendConfigResolver::new;
     private ConfigVersionManager configVersionManager;
+    private Supplier<ProcessorFactory> processorFactorySupplier;
 
     private ConciergeHub() {
 
@@ -44,12 +44,6 @@ public class ConciergeHub {
         return this;
     }
 
-    public ConciergeHub withVersionSupplier(Supplier<VersionGenerator> versionSupplier) {
-        Preconditions.checkNotNull(versionSupplier, "Version ID Generator can't be null");
-        this.versionSupplier = versionSupplier;
-        return this;
-    }
-
     public ConciergeHub withAppendConfigResolverSupplier(Supplier<AppendConfigResolver> appendConfigResolver) {
         Preconditions.checkNotNull(appendConfigResolver, "Append Config Resolver can't be null");
         this.appendConfigResolverSupplier = appendConfigResolver;
@@ -62,6 +56,12 @@ public class ConciergeHub {
         return this;
     }
 
+    public ConciergeHub withProcessorFactory(Supplier<ProcessorFactory> pfSupplier) {
+        Preconditions.checkNotNull(pfSupplier, "Processor Factory can't be null");
+        this.processorFactorySupplier = pfSupplier;
+        return this;
+    }
+
     public ConciergeHub build() {
         Arrays.stream(ConfigEvent.values()).forEach(this::buildProcessor);
         return this;
@@ -71,32 +71,52 @@ public class ConciergeHub {
         processors.putIfAbsent(configEvent, configEvent.accept(new ConfigEventVisitor<>() {
             @Override
             public ConciergeProcessor configCreate() {
-                return new CreateConfigProcessor(repositorySupplier, versionSupplier);
+                return CreateConfigProcessor.builder()
+                        .processorFactory(processorFactorySupplier)
+                        .repositorySupplier(repositorySupplier)
+                        .build();
             }
 
             @Override
             public ConciergeProcessor configAppend() {
-                return new AppendConfigProcessor(repositorySupplier, appendConfigResolverSupplier);
+                return AppendConfigProcessor.builder()
+                        .repositorySupplier(repositorySupplier)
+                        .appendConfigResolverSupplier(appendConfigResolverSupplier)
+                        .processorFactory(processorFactorySupplier)
+                        .build();
             }
 
             @Override
             public ConciergeProcessor configUpdate() {
-                return new UpdateConfigProcessor(repositorySupplier);
+                return UpdateConfigProcessor.builder()
+                        .repositorySupplier(repositorySupplier)
+                        .processorFactory(processorFactorySupplier)
+                        .build();
             }
 
             @Override
             public ConciergeProcessor configApprove() {
-                return new ApproveConfigProcessor(repositorySupplier);
+                return ApproveConfigProcessor.builder()
+                        .repositorySupplier(repositorySupplier)
+                        .processorFactory(processorFactorySupplier)
+                        .build();
             }
 
             @Override
             public ConciergeProcessor configReject() {
-                return new RejectConfigProcessor(repositorySupplier);
+                return RejectConfigProcessor.builder()
+                        .processorFactory(processorFactorySupplier)
+                        .repositorySupplier(repositorySupplier)
+                        .build();
             }
 
             @Override
             public ConciergeProcessor configActivate() {
-                return new ActivateConfigProcessor(repositorySupplier, configVersionManager);
+                return ActivateConfigProcessor.builder()
+                        .repositorySupplier(repositorySupplier)
+                        .configVersionManager(configVersionManager)
+                        .processorFactory(processorFactorySupplier)
+                        .build();
             }
         }));
     }