+++ concierge-elastic/src/main/java/com/grookage/concierge/elastic/repository/ElasticRepository.java
@@ -37,10 +37,7 @@ import com.grookage.concierge.repository.ConciergeRepository;
 import lombok.Getter;
 import lombok.SneakyThrows;
 
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
@@ -79,7 +76,7 @@ public class ElasticRepository implements ConciergeRepository {
     }
 
     /* Fields and values are being lower-cased, before adding as clauses, since elasticsearch deals with lowercase only */
-    private List<FieldValue> getNormalizedValues(Set<String> terms) {
+    private List<FieldValue> getNormalizedValues(Collection<String> terms) {
         return terms.stream().map(FieldValue::of).toList();
     }
 
@@ -157,40 +154,17 @@ public class ElasticRepository implements ConciergeRepository {
     }
 
     @Override
-    public List<ConfigDetails> getStoredRecords(String namespace,
+    public List<ConfigDetails> getStoredRecords(Set<String> namespaces,
                                                 Set<String> configNames,
                                                 Set<ConfigState> configStates) {
-        final var namespaceQuery = TermQuery.of(p -> p.field(NAMESPACE).value(namespace))._toQuery();
-        final var configQuery = TermsQuery.of(q -> q.field(CONFIG_NAME)
-                .terms(t -> t.value(getNormalizedValues(configNames))))._toQuery();
-        final var configStateQuery = TermsQuery.of(q -> q.field(CONFIG_STATE)
-                        .terms(t -> t.value(getNormalizedValues(configStates.stream().map(Enum::name).collect(Collectors.toSet()))))).
-                _toQuery();
+        final var namespaceQuery = namespaces.isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
+                TermsQuery.of(q -> q.field(NAMESPACE).terms(t -> t.value(getNormalizedValues(namespaces))))._toQuery();
+        final var configQuery = configNames.isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
+                TermsQuery.of(q -> q.field(CONFIG_NAME).terms(t -> t.value(getNormalizedValues(configNames))))._toQuery();
+        final var configStateQuery = configStates.isEmpty() ? MatchAllQuery.of(q -> q)._toQuery() :
+                TermsQuery.of(q -> q.field(CONFIG_STATE).terms(t -> t.value(getNormalizedValues(configStates.stream().map(Enum::name).collect(Collectors.toSet()))))).
+                        _toQuery();
         final var searchQuery = BoolQuery.of(q -> q.must(List.of(namespaceQuery, configQuery, configStateQuery)))._toQuery();
-        return queryDetails(searchQuery, storedElasticRecordHit -> true);
-    }
-
-    @Override
-    public List<ConfigDetails> getActiveStoredRecords(Set<String> namespaces) {
-        final var namespaceQuery = namespaces.isEmpty() ?
-                MatchAllQuery.of(q -> q)._toQuery() :
-                TermsQuery.of(q -> q.field(NAMESPACE)
-                        .terms(t -> t.value(getNormalizedValues(namespaces))))._toQuery();
-        final var searchQuery = BoolQuery.of(q -> q.must(List.of(namespaceQuery)))._toQuery();
-        return queryDetails(searchQuery, storedElasticRecordHit -> storedElasticRecordHit.source() != null &&
-                storedElasticRecordHit.source().getConfigState() == ConfigState.ACTIVATED);
-    }
-
-    @Override
-    @SneakyThrows
-    public List<ConfigDetails> getStoredRecords(Set<String> namespaces) {
-        if (namespaces.isEmpty()) {
-            return getStoredRecords();
-        }
-
-        final var namespaceQuery = TermsQuery.of(q -> q.field(NAMESPACE)
-                .terms(t -> t.value(getNormalizedValues(namespaces))))._toQuery();
-        final var searchQuery = BoolQuery.of(q -> q.must(List.of(namespaceQuery)))._toQuery();
         return queryDetails(searchQuery, storedElasticRecordHit -> true);
     }