+++ concierge-models/src/main/java/com/grookage/concierge/models/config/ConfigDetails.java
@@ -16,7 +16,7 @@ import java.util.Set;
 @Builder
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ConfigDetails implements Comparable<ConfigDetails> {
+public class ConfigDetails {
     @NotNull
     ConfigKey configKey;
     @NotNull
@@ -39,10 +39,5 @@ public class ConfigDetails implements Comparable<ConfigDetails> {
         }
 
         configHistories.add(configHistoryItem);
-    }
-
-    @Override
-    public int compareTo(ConfigDetails o) {
-        return configKey.compareTo(o.getConfigKey());
     }
 }