+++ concierge-models/src/main/java/com/grookage/concierge/models/config/ConfigKey.java
@@ -16,7 +16,7 @@ import java.util.Locale;
 @Builder
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ConfigKey implements Comparable<ConfigKey> {
+public class ConfigKey {
 
     @NotBlank
     String namespace;
@@ -24,21 +24,12 @@ public class ConfigKey implements Comparable<ConfigKey> {
     String configName;
     @NotBlank
     String version;
-    ConfigType configType;
+    @NotBlank
+    String configType;
 
     @JsonIgnore
     public String getReferenceId() {
         return Joiner.on(".").join(namespace, configName, version).toUpperCase(Locale.ROOT);
-    }
-
-    @JsonIgnore
-    public String getReferenceTag() {
-        return Joiner.on(".").join(namespace, configName).toUpperCase(Locale.ROOT);
-    }
-
-    @JsonIgnore
-    public boolean latest() {
-        return null != version && version.equalsIgnoreCase("latest");
     }
 
     @Override
@@ -58,8 +49,4 @@ public class ConfigKey implements Comparable<ConfigKey> {
         return (thatKey.getReferenceId().equals(this.getReferenceId()));
     }
 
-    @Override
-    public int compareTo(ConfigKey o) {
-        return this.version.compareTo(o.getVersion());
-    }
 }