+++ concierge-models/src/main/java/com/grookage/concierge/models/config/ConfigType.java
@@ -1,6 +1,0 @@
-package com.grookage.concierge.models.config;
-
-public interface ConfigType {
-
-    String name();
-}