+++ concierge-models/src/main/java/com/grookage/concierge/models/ingestion/ConfigurationRequest.java
@@ -3,7 +3,6 @@ package com.grookage.concierge.models.ingestion;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.grookage.concierge.models.config.ConfigKey;
-import com.grookage.concierge.models.config.ConfigType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -22,10 +21,13 @@ public class ConfigurationRequest {
     String namespace;
     @NotBlank
     String configName;
+    @NotBlank
+    String versionId;
     String description;
     @NotNull
     Object data;
-    ConfigType configType;
+    @NotBlank
+    String configType;
     String message;
 
     @JsonIgnore
@@ -34,7 +36,7 @@ public class ConfigurationRequest {
                 .namespace(namespace)
                 .configName(configName)
                 .configType(configType)
-                .version("latest")
+                .version(versionId)
                 .build();
     }
 }